admin: move vitest tests to tests/vitest
